refactor: don't let aio_provider.h in include
