+++ include/dsn/tool_api.h
@@ -61,7 +61,6 @@ Component providers define the interface for the local components (e.g., network
 #include <dsn/tool-api/message_parser.h>
 #include <dsn/tool-api/logging_provider.h>
 #include <dsn/tool-api/timer_service.h>
-#include <dsn/tool-api/aio_provider.h>
 
 namespace dsn {
 namespace tools {
@@ -138,8 +137,6 @@ DSN_API bool register_component_provider(const char *name,
                                          ::dsn::provider_type type);
 DSN_API bool
 register_component_provider(const char *name, network::factory f, ::dsn::provider_type type);
-DSN_API bool
-register_component_provider(const char *name, aio_provider::factory f, ::dsn::provider_type type);
 DSN_API bool
 register_component_provider(const char *name, env_provider::factory f, ::dsn::provider_type type);
 DSN_API bool register_component_provider(const char *name,
