+++ src/core/aio/disk_engine.cpp
@@ -330,13 +330,4 @@ void disk_engine::complete_io(aio_task *aio, error_code err, uint32_t bytes, int
         }
     }
 }
-
-namespace tools {
-namespace internal_use_only {
-bool register_component_provider(const char *name, aio_provider::factory f, dsn::provider_type type)
-{
-    return dsn::utils::factory_store<aio_provider>::register_factory(name, f, type);
-}
-} // namespace internal_use_only
-} // namespace tools
 } // namespace dsn
