+++ src/core/aio/native_linux_aio_provider.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include <dsn/tool-api/aio_provider.h>
+#include "aio_provider.h"
 
 #include <dsn/tool_api.h>
 #include <dsn/utility/synchronize.h>
